requirements.txt
python-telegram-bot>=20.0
yfinance
apscheduler
pytz
